+++ docs/docs/conf.py
@@ -8,4 +8,18 @@ exclude_patterns = ["out/**"]
 html_favicon = None
 
 site_url = 'https://crate.io/docs/croud/en/latest/'
-extensions = ['sphinx_sitemap']
+extensions = ['sphinx_sitemap', 'sphinxarg.ext']
+
+# The file extension
+source_suffix = '.rst'
+
+# The master toctree document.
+master_doc = 'index'
+
+
+html_static_path = ['_static']
+html_context = {
+    'css_files': [
+        '_static/sphinxarg.css',  # overrides
+    ],
+}