+++ CHANGES.rst
@@ -5,6 +5,12 @@ Changes for croud
 Unreleased
 ==========
 
+0.11.0 - 2019/04/17
+===================
+
+- Added the commands ``monitoring grafana`` that can enable and
+  disable Grafana dashboards for a project.
+
 0.10.0 - 2019/04/01
 ===================